# OpenAPI.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Ski-Service API",
		"description": "Ski Service Anmeldung",
		"termsOfService": "https://www.ipso.ch/ibz",
		"contact": {
			"name": "INFEFZ",
			"url": "https://www.ipso.ch/ibz"
		},
		"license": {
			"name": "IPSO",
			"url": "https://www.ipso.ch/ibz"
		},
		"version": "v1"
	},
	"paths": {
		"/Registration": {
			"get": {
				"tags": [
					"Registration"
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Registration"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Registration"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Registration"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"Registration"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Registration"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/Registration"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/Registration"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
		"/Registration/{id}": {
			"get": {
				"tags": [
					"Registration"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/Registration"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Registration"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/Registration"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"Registration"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Registration"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/Registration"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/Registration"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"delete": {
				"tags": [
					"Registration"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Registration": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"priority": {
						"type": "string",
						"nullable": true
					},
					"service": {
						"type": "string",
						"nullable": true
					},
					"create_date": {
						"type": "string",
						"format": "date-time"
					},
					"pickup_date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"additionalProperties": false
			}
		}
	}
}